requirements.txt
python 3.12
torch
torchvision.io
pandas
clearml
